directory hierarchy refactor
