+++ ContactBook/app/src/main/res/layout/fragment_contact_detail.xml
@@ -25,7 +25,8 @@
         app:cardCornerRadius="20dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
+        app:layout_constraintStart_toStartOf="parent"
+        >
 
         <androidx.appcompat.widget.LinearLayoutCompat
             android:layout_width="match_parent"
@@ -39,6 +40,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:background="@drawable/bg_rounded_tv"
+                android:layout_marginTop="20dp"
                 android:paddingStart="10dp"
                 android:text="Contact name"
                 android:textColor="@color/black"
